checks if yml exists first

+++ lib/generators/export/templates/export.rb
@@ -1,6 +1,6 @@
 # Loads export.yml and sets key=value to the local environment
 
-config = YAML.load(File.open("#{Rails.root}/config/export.yml"))
-unless Rails.env == 'production'
+if File.exists? "#{Rails.root}/config/export.yml"
+  config = YAML.load(File.open("#{Rails.root}/config/export.yml"))
   config.each_pair { |key,value| ENV[key] = value }
 end
